requirements.txt
django==4.0.6

djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-ckeditor==6.4.2